perf(events): move controller key events to BaseMenu
